Runtime restructuring

+++ src/engine/renderer/Window.h
@@ -1,29 +1,54 @@
 #pragma once
-#include <iostream>
 #define TS_RENDERER_RAYLIB
 #ifdef TS_RENDERER_RAYLIB
-#include <raylib.h>
-#include <rlgl.h>
 
 #endif
 
 namespace Techstorm {
+	/// <summary>
+	///
+	/// </summary>
+	/// @deprecated Do not use this class as it will be removed.
 	class Window {
 	public:
 
+		/// <summary>
+		/// Finalizes an instance of the <see cref="Window"/> class.
+		/// </summary>
+		/// @deprecated See class deprecation warning
 		~Window();
 
-		// TODO : Make this the constructor
-		int initialize(uint16_t width = 600, uint16_t height = 800, const char* title = "Untitled Techstorm Project", unsigned int windowConfig = 0);
+		/// <summary>
+		/// Initializes the specified width, height, title, and window configuration flags.
+		/// </summary>
+		/// <param name="width">The width.</param>
+		/// <param name="height">The height.</param>
+		/// <param name="title">The title.</param>
+		/// <param name="windowConfig">The window configuration.</param>
+		/// <returns></returns>
+		/// @deprecated See class deprecation warning
+		int initialize(int width = 600, int height = 800, const char* title = "Untitled Techstorm Project", unsigned int windowConfig = 0);
+
+		/// <summary>
+		/// Should the window close.
+		/// </summary>
+		/// <returns>The value of WindowShouldClose()</returns>
+		/// @deprecated See class deprecation warning
 		bool shouldClose();
+
+		/// <summary>
+		/// Determines whether this instance is ready.
+		/// </summary>
+		/// <returns>
+		///   <c>true</c> if this instance is ready; otherwise, <c>false</c>.
+		/// </returns>
+		/// @deprecated See class deprecation warning
 		bool isReady();
 
-		
 	private:
 
-	
-		uint16_t mWidth = 600;
-		uint16_t mHeight = 800;
+		int mWidth = 600;
+		int mHeight = 800;
 		const char* mTitle = "Untitled Techstorm Project";
 		unsigned int mWindowConfig;
 	};
